Create VarTypes
Instance Var : is declared inside class but outside body of method; not declared as static; value is instance specific; not shared among instances.
Local Variable : is declared inside the body of the method. Can use this only within that method.
Static Variable : is declared as static. It cannot be local. You can create a single copy of static variable and share among all the instances of the class. Memory allocation for static variable happens only once when the class is loaded in the memory